requirements.txt
discord-rewrite>=1.0.0a
